very dirty hack addressing cannon tower pitch

+++ trunk/Model/Terrain/AntinTerrainGenerator.cpp
@@ -24,7 +24,7 @@ AntinTerrainGenerator::~AntinTerrainGenerator()
 
 void AntinTerrainGenerator::generateHeightmap(unsigned char** ppVertexHeightData, const unsigned short terrainSize)
 {
-    if(m_Seed == 0) srand(time(0));
+    if(m_Seed == 0) srand((unsigned int)time(0));
     else srand(m_Seed);
 
     //PERUNAPELTOGENERAATTORI aka. just testing algorithms - NOT an actual way to make good terrain!
